sat_mqtt/requirements.txt
paho-mqtt==1.6.1
ortools==9.9.3963
